Addressing issues and bug fixes.
Checked for nan values in group_channels and updated it to raise an
exception if streams with more than 3 channels are created.
Undefined hdr variable replaced with stats in smc reader removed.
Dataframe creation is possible by directory or set of streams.
Lat/lon bug in streams_to_dataframe fixed.